lesson3 1) Получение и обработка данных в отдельных классах 2) Ввод города в диалоговом окне
