+++ app/src/main/java/com/example/lesson1/CityList.java
@@ -1,19 +1,16 @@
 package com.example.lesson1;
 
-import android.content.res.Configuration;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.AutoCompleteTextView;
-import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentTransaction;
+
+import com.example.lesson1.fragments.ShowWeatherFragment;
 
 import static android.view.Gravity.CENTER;
