+++ app/src/main/res/layout/fragment_dialog.xml
@@ -6,17 +6,25 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content">
+    android:layout_height="match_parent">
 
-    <ImageButton
-        android:id="@+id/button3"
-        android:layout_width="50dp"
+    <AutoCompleteTextView
+        android:id="@+id/autoCompleteTextViewD"
+        android:layout_width="200dp"
         android:layout_height="50dp"
-        android:src="@drawable/change_city"
-        app:layout_constraintEnd_toEndOf="parent"
+        android:gravity="bottom"
+        android:hint="@string/enter_city"
+        app:layout_constraintEnd_toStartOf="@+id/imageButton"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-
+    <ImageButton
+        android:id="@+id/imageButton"
+        android:layout_width="50dp"
+        android:layout_height="50dp"
+        android:src="@drawable/magnifier_2"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
