+++ app/src/main/res/values/strings.xml
@@ -110,5 +110,7 @@
         <item>Without precipitation</item>
         <item>Without precipitation</item>
     </string-array>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
